refactor : Main 클래스 이름 변경 및 InputView, OutputView view 패키지로 이동
